Unify army assignment into TerritoryAssignmentService
